Use phpsass extension with fallback to scssphp.

+++ includes/pb-sass.php
@@ -14,19 +14,26 @@ namespace PressBooks\SASS;
  *
  * @return string the compiled CSS
  */
-function compile( $scss, $options = array() ) {
+function compile( $scss, $includes = array() ) {
 	
 	$css = '/* Silence is golden. */'; // If no SCSS input was passed, prevent file write errors by putting a comment in the CSS output.
 
 	if ( $scss !== '' ) {
-		$scss_file = array_search( 'uri', @array_flip( stream_get_meta_data( $GLOBALS[mt_rand()] = tmpfile() ) ) );
-		rename( $scss_file, $scss_file .= '.scss' ); 
-		register_shutdown_function( create_function( '', "unlink('{$scss_file}');" ) ); 
-		file_put_contents( $scss_file, $scss );
-		require_once( PB_PLUGIN_DIR . 'symbionts/phpsass/SassLoader.php' );
-		
-		$sass = new \SassParser( $options );
-		$css = $sass->toCss( $scss_file );
+		if ( extension_loaded( 'sass' ) ) { // use sassphp extension
+			$scss_file = array_search( 'uri', @array_flip( stream_get_meta_data( $GLOBALS[mt_rand()] = tmpfile() ) ) );
+			rename( $scss_file, $scss_file .= '.scss' ); 
+			register_shutdown_function( create_function( '', "unlink('{$scss_file}');" ) ); 
+			file_put_contents( $scss_file, $scss );
+			$sass = new \Sass();
+			$include_paths = implode( ':', $includes );
+			$sass->setIncludePath( $include_paths );
+			$css = $sass->compileFile( $scss_file );
+		} else { // use scssphp library
+			require_once( PB_PLUGIN_DIR . 'symbionts/scssphp/scss.inc.php' );
+			$sass = new \Leafo\ScssPhp\Compiler;
+			$sass->setImportPaths( $includes );
+			$css = $sass->compile( $scss );
+ 		}
 	}
 	
 	return $css;
